+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [7.6.6.dev0 (unreleased)](https://github.com/kdeldycke/click-extra/compare/v7.6.5...main)
+## [7.7.0.dev0 (unreleased)](https://github.com/kdeldycke/click-extra/compare/v7.6.5...main)
 
 > [!WARNING]
 > This version is **not released yet** and is under active development.